Moved the constants code into Helpers/constants.jsx
